rename: 컴포넌트 파일 경로 단축
