+++ src/pages/Payment/components/Agreement.tsx
