+++ src/pages/Payment/components/PaymentAmount.tsx
