+++ src/pages/Payment/index.tsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import { TossPaymentsWidgets } from '@tosspayments/tosspayments-sdk';
 import clsx from 'clsx';
 import Agreement from './components/Agreement';
-import ContactModal from './components/ContactModal/ContactModal';
+import ContactModal from './components/ContactModal';
 import PaymentAmount from './components/PaymentAmount';
-import StoreRequestModal from './components/StoreRequestModal/StoreRequestModal';
+import StoreRequestModal from './components/StoreRequestModal';
 import TossWidget from './components/TossWidget';
 import Bike from '@/assets/Main/agriculture.svg';
 import Home from '@/assets/Main/home-icon.svg';
